fix(data): organize data management Fixes #46
